refactor(discord): split embed/config/sender/event along natural seams
Each oversized module (>700-line hard cap, CLAUDE.md "File organization")
becomes a directory with a re-exporting mod.rs:

- event.rs (742) -> event/{channel_kind,event_kind,payload}.rs
- embed.rs (1182) -> embed/{builder,format,budget}.rs
- config.rs (1067) -> config/{model,toggles,parse,watcher}.rs
- sender.rs (845) -> sender/{handle,stats,token_bucket,task,http,mock}.rs

Pure code movement: no logic, signature, or behaviour changes. Each new
mod.rs re-exports its submodules' public items so every external
`use cimmeria_discord::...` path stays identical (zero public-surface
change). Inline `#[cfg(test)] mod tests` moved to the submodule they
exercise; only import paths changed, never test logic. All 44 existing
tests preserved verbatim (verified by name diff vs HEAD), including the
`event_kind_all_matches_variant_count` and `every_event_kind_routes`
pinning tests.

Net-new test: DisconnectReason::as_str variant pin (previously only
from_label was covered).

Part of #529.
