+++ crates/discord/src/event/payload.rs
@@ -1,233 +1,12 @@
-//! Typed event taxonomy. Every server-side notification is one of these
-//! variants — the type system makes it impossible to construct an event
-//! with the wrong payload shape, and `EventKind` provides a discriminant
-//! that lines up with the per-event toggles in [`crate::config`].
-//!
-//! Events route to one of 8 channels (`ChannelKind`). The mapping lives in
-//! [`crate::router`] and is exhaustively tested.
+//! The rich [`Event`] payload, its supporting enums, and the
+//! discriminant/severity derivations.
 
 use std::net::SocketAddr;
 
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
-// ── ChannelKind: logical destination ────────────────────────────────────
-
-/// Logical channel a category of events posts to. Each variant maps 1:1 to
-/// a webhook URL in `Config::channels`. Eight channels keeps the signal-to-
-/// noise tiers separable while staying inside Discord's per-webhook rate
-/// limit (5 msg / 2 s) for each.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ChannelKind {
-    /// Server up/down/panic — low volume, high signal.
-    Lifecycle,
-    /// Login, logout, auth fail, unexpected disconnect.
-    Auth,
-    /// World entry / exit.
-    World,
-    /// Player chat (global/say/whisper/guild/team) — high volume.
-    Chat,
-    /// Level up, death, respawn, mission events, loot.
-    Gameplay,
-    /// GM-issued admin commands (teleport, spawn, item grant, ...).
-    Gm,
-    /// `warn!` / `error!` / wire-format / DB / assertion / mercury timeout.
-    Errors,
-    /// Latency, packet loss, memory, tick stall, AoI burst, outbox lag.
-    Ops,
-}
-
-impl ChannelKind {
-    /// All eight channels in a stable order. Used by config validation +
-    /// stats reporting.
-    pub const ALL: &'static [Self] = &[
-        Self::Lifecycle,
-        Self::Auth,
-        Self::World,
-        Self::Chat,
-        Self::Gameplay,
-        Self::Gm,
-        Self::Errors,
-        Self::Ops,
-    ];
-
-    /// Lowercase short name, matches the TOML key in `[discord.channels]`.
-    pub const fn as_str(self) -> &'static str {
-        match self {
-            Self::Lifecycle => "lifecycle",
-            Self::Auth => "auth",
-            Self::World => "world",
-            Self::Chat => "chat",
-            Self::Gameplay => "gameplay",
-            Self::Gm => "gm",
-            Self::Errors => "errors",
-            Self::Ops => "ops",
-        }
-    }
-}
-
-// ── EventKind: 28-variant discriminant ──────────────────────────────────
-
-/// Discriminant of [`Event`] — used as the key for the per-event toggle map
-/// in [`crate::config::EventToggles`] and as the value type for the router.
-///
-/// **Adding a variant requires touching three places**: this enum, the
-/// matching `Event` variant, and `EventToggles` field. The router test
-/// `every_event_kind_has_a_channel` panics on a `match` `_` arm so a
-/// missing route fails at compile time.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum EventKind {
-    // Lifecycle
-    ServerStartup,
-    ServerShutdown,
-    ServerPanic,
-
-    // Auth
-    PlayerLogin,
-    PlayerLogout,
-    PlayerDisconnect,
-    PlayerAuthFailed,
-
-    // World
-    PlayerWorldEntry,
-    PlayerWorldExit,
-
-    // Chat
-    ChatGlobal,
-    ChatSay,
-    ChatWhisper,
-    ChatGuild,
-    ChatTeam,
-    ChatCommand,
-
-    // Gameplay
-    PlayerLevelUp,
-    PlayerDeath,
-    PlayerRespawn,
-    MissionAccepted,
-    MissionCompleted,
-    MissionFailed,
-    MissionRewardGranted,
-    LootGenerated,
-    ItemUsed,
-
-    // GM
-    GmCommand,
-    GmTeleport,
-    GmSpawn,
-    GmItemGrant,
-
-    // Errors
-    Warning,
-    Error,
-    WireFormatError,
-    DbError,
-    AssertionFailure,
-    MercuryTimeout,
-
-    // Ops
-    HighLatency,
-    PacketLossSpike,
-    MemoryWarning,
-    TickStall,
-    AoiBurstWarning,
-    OutboxLag,
-}
-
-impl EventKind {
-    /// All variants in declaration order — used by config validation to
-    /// ensure every variant has a toggle default, and by docs generation.
-    pub const ALL: &'static [Self] = &[
-        Self::ServerStartup,
-        Self::ServerShutdown,
-        Self::ServerPanic,
-        Self::PlayerLogin,
-        Self::PlayerLogout,
-        Self::PlayerDisconnect,
-        Self::PlayerAuthFailed,
-        Self::PlayerWorldEntry,
-        Self::PlayerWorldExit,
-        Self::ChatGlobal,
-        Self::ChatSay,
-        Self::ChatWhisper,
-        Self::ChatGuild,
-        Self::ChatTeam,
-        Self::ChatCommand,
-        Self::PlayerLevelUp,
-        Self::PlayerDeath,
-        Self::PlayerRespawn,
-        Self::MissionAccepted,
-        Self::MissionCompleted,
-        Self::MissionFailed,
-        Self::MissionRewardGranted,
-        Self::LootGenerated,
-        Self::ItemUsed,
-        Self::GmCommand,
-        Self::GmTeleport,
-        Self::GmSpawn,
-        Self::GmItemGrant,
-        Self::Warning,
-        Self::Error,
-        Self::WireFormatError,
-        Self::DbError,
-        Self::AssertionFailure,
-        Self::MercuryTimeout,
-        Self::HighLatency,
-        Self::PacketLossSpike,
-        Self::MemoryWarning,
-        Self::TickStall,
-        Self::AoiBurstWarning,
-        Self::OutboxLag,
-    ];
-
-    /// Lowercase snake_case name. Matches the TOML key in `[discord.events]`.
-    pub const fn as_str(self) -> &'static str {
-        match self {
-            Self::ServerStartup => "server_startup",
-            Self::ServerShutdown => "server_shutdown",
-            Self::ServerPanic => "server_panic",
-            Self::PlayerLogin => "player_login",
-            Self::PlayerLogout => "player_logout",
-            Self::PlayerDisconnect => "player_disconnect",
-            Self::PlayerAuthFailed => "player_auth_failed",
-            Self::PlayerWorldEntry => "player_world_entry",
-            Self::PlayerWorldExit => "player_world_exit",
-            Self::ChatGlobal => "chat_global",
-            Self::ChatSay => "chat_say",
-            Self::ChatWhisper => "chat_whisper",
-            Self::ChatGuild => "chat_guild",
-            Self::ChatTeam => "chat_team",
-            Self::ChatCommand => "chat_command",
-            Self::PlayerLevelUp => "player_level_up",
-            Self::PlayerDeath => "player_death",
-            Self::PlayerRespawn => "player_respawn",
-            Self::MissionAccepted => "mission_accepted",
-            Self::MissionCompleted => "mission_completed",
-            Self::MissionFailed => "mission_failed",
-            Self::MissionRewardGranted => "mission_reward_granted",
-            Self::LootGenerated => "loot_generated",
-            Self::ItemUsed => "item_used",
-            Self::GmCommand => "gm_command",
-            Self::GmTeleport => "gm_teleport",
-            Self::GmSpawn => "gm_spawn",
-            Self::GmItemGrant => "gm_item_grant",
-            Self::Warning => "warning",
-            Self::Error => "error",
-            Self::WireFormatError => "wire_format_error",
-            Self::DbError => "db_error",
-            Self::AssertionFailure => "assertion_failure",
-            Self::MercuryTimeout => "mercury_timeout",
-            Self::HighLatency => "high_latency",
-            Self::PacketLossSpike => "packet_loss_spike",
-            Self::MemoryWarning => "memory_warning",
-            Self::TickStall => "tick_stall",
-            Self::AoiBurstWarning => "aoi_burst_warning",
-            Self::OutboxLag => "outbox_lag",
-        }
-    }
-}
+use super::EventKind;
 
 // ── Severity: drives embed color ────────────────────────────────────────
 
@@ -249,7 +28,8 @@ pub enum Severity {
 // ── Event: the rich payload ─────────────────────────────────────────────
 
 /// A single Discord-bound event. The variant uniquely determines an
-/// [`EventKind`], a [`ChannelKind`], and a [`Severity`].
+/// [`EventKind`], a [`ChannelKind`](super::ChannelKind), and a
+/// [`Severity`].
 ///
 /// **Field types are deliberately small and well-defined.** No `dyn Any` /
 /// `serde_json::Value` payloads — every field is typed so the embed builder
@@ -669,32 +449,6 @@ impl Event {
 mod tests {
     use super::*;
 
-    /// Pin that every `EventKind` variant has a unique `as_str`. A
-    /// duplicate would let two variants compete for the same TOML toggle
-    /// key.
-    #[test]
-    fn every_event_kind_has_unique_snake_case_name() {
-        let mut seen = std::collections::HashSet::new();
-        for k in EventKind::ALL {
-            assert!(
-                seen.insert(k.as_str()),
-                "duplicate snake-case name: {}",
-                k.as_str()
-            );
-        }
-        assert_eq!(seen.len(), EventKind::ALL.len(), "ALL has duplicates");
-    }
-
-    /// Pin that every `ChannelKind` variant has a unique `as_str`.
-    #[test]
-    fn every_channel_kind_has_unique_snake_case_name() {
-        let mut seen = std::collections::HashSet::new();
-        for c in ChannelKind::ALL {
-            assert!(seen.insert(c.as_str()));
-        }
-        assert_eq!(seen.len(), ChannelKind::ALL.len());
-    }
-
     /// `DisconnectReason::from_label` is the single mapping the auth-channel
     /// `player_disconnect` seam relies on. Pin every label the
     /// `destroy_client_entities` choke point stamps, plus the unknown-label
@@ -729,14 +483,19 @@ mod tests {
         );
     }
 
-    /// `EventKind::ALL` length is the canonical count of toggleable event
-    /// types. Adding a variant without updating `ALL` would silently break
-    /// config-default generation.
+    /// `DisconnectReason::as_str` is the stable snake_case label each
+    /// variant serialises to. Pin every variant so a rename (which would
+    /// silently change a downstream log/metric label) trips here. Net-new
+    /// coverage added with the event-module split — the previous suite
+    /// only exercised `from_label`, leaving `as_str` unguarded.
     #[test]
-    fn event_kind_all_matches_variant_count() {
-        // Update this when adding/removing an EventKind variant. The
-        // doc-comment on EventKind enumerates the three places to touch
-        // (event.rs, EventToggles, router).
-        assert_eq!(EventKind::ALL.len(), 40);
+    fn disconnect_reason_as_str_is_stable() {
+        assert_eq!(DisconnectReason::Clean.as_str(), "clean");
+        assert_eq!(DisconnectReason::Timeout.as_str(), "timeout");
+        assert_eq!(DisconnectReason::PeerReset.as_str(), "peer_reset");
+        assert_eq!(
+            DisconnectReason::ServerInitiated.as_str(),
+            "server_initiated"
+        );
     }
 }
